Replace SOR authorization with user ID for integration
This update modifies the Integration Controller to use a user ID instead of the previous authorization flag. Functions to get and set the user ID are added in utils.js. Moreover, upgrades were made to handle the user ID in cases of revoking authorization, improving overall handling and storage of user data.