Update sysadmin handling in scripts
The handling of sysadmin created status was added to constants and utils.js scripts. This change aims to improve error management in the case of sysadmin creation particularly in the addMemberToGroupMapping script, where the sysadmin status is checked prior to execution. If sysadmin does not exist or an error occurs during sysadmin assignment, the sysadmin is recreated.