Add data loss recovery migration in GroupsController
The update introduces a function in the GroupsController that is activated in the event of total data loss recovery. It performs a fetch command to "/doMigration". As part of the process, this change enhances the reliability and resilience of the wallet DBEnclave by ensuring data can be restored following complete loss.